refactor(757): split validation.py into select / inspect / evaluate
Second step of the #757 workflow-namespace migration. Physically decomposes the
171 KB catch-all validation.py (three interleaved concerns) into three
implementation modules aligned to the workflow, with validation.py kept as a
compatibility shim.

- select.py  — K/model selection: search_k, SearchKResult, select_model,
  quality_frontier, plot_*, BestKExplanation (+ SEARCH_K_DIRECTIONS and the
  frontier/criterion helpers).
- inspect.py — reading fitted topics: label_topics, topic_table, frex, mmr,
  relevance, find_thoughts(_html), topics_for_term, topic_correlation,
  prepare_pyldavis (leaf module; imports nothing from its siblings).
- evaluate.py — credibility diagnostics: diagnostics, perplexity, heldout,
  residuals, align_topics, topic_stability, bootstrap_stability, flag_topics,
  topic_dendrogram.

The concern DAG is acyclic (inspect <- evaluate <- select): evaluate imports
`frex` from inspect; select imports `Heldout`/`perplexity`/`check_residuals`/
`eval_heldout`/`_kl` from evaluate. Partitioned by AST so the moved source is
byte-preserved.

validation.py is now a thin shim that re-exports all 68 names (public + the
private helpers other modules import: `_hungarian`, `_accepts_kwarg`,
`_classify_alignment`, `align_topics`, ...), so `topica.validation.X`,
`from topica.validation import X`, and every `topica.X` root alias keep working
unchanged. No internal import repointing needed this PR; that cleanup lands with
the namespace/docs work.

Verification: full suite 4959 passed / 38 skipped; preflight + mkdocs --strict
clean; import identities (`topica.X is topica.<ns>.X is validation.X`) verified;
stdlib `inspect` not shadowed inside the package.
